done before deployment

+++ app/actions/deleteRecipe.js
@@ -1,24 +1,25 @@
 'use server';
 
 import Recipe from '@/models/Recipe';
+import { getSessionUser } from '@/utils/getSessionUser';
 import { revalidatePath } from 'next/cache';
 
-// import { promises as fs } from 'fs';
-
 const deleteRecipe = async (recipeId) => {
+  const sessionUser = await getSessionUser();
+
+  if (!sessionUser || !sessionUser.userId) {
+    throw new Error('User ID is required');
+  }
+
+  const { userId } = sessionUser;
   const recipe = await Recipe.findById(recipeId);
-
+  if (!recipe) {
+    throw new Error('Recipe not found');
+  }
+  if (recipe.owner.toString() !== userId) {
+    throw new Error('Unauthorized');
+  }
   await recipe.deleteOne();
-
-  // const filePath = 'recipes.json';
-
-  // const jsonData = await fs.readFile(filePath, 'utf-8');
-  // const data = JSON.parse(jsonData);
-
-  // const updatedData = data.filter((recipe) => recipe._id !== recipeId);
-
-  // const updatedJsonData = JSON.stringify(updatedData, null, 2);
-  // await fs.writeFile(filePath, updatedJsonData);
 
   revalidatePath('/', 'layout');
 };
